move ajax that was separated out into views.py
