--- commcare_connect/solicitations/fixtures/real_eois.json
@@ -1,47 +1,0 @@
-[
-  {
-    "model": "organization.organization",
-    "pk": 1,
-    "fields": {
-      "name": "Dimagi",
-      "slug": "dimagi",
-      "program_manager": true,
-      "created_by": "system",
-      "date_created": "2024-01-01T00:00:00Z"
-    }
-  },
-  {
-    "model": "program.program",
-    "pk": 1,
-    "fields": {
-      "name": "CommCare Connect Child Health Campaign",
-      "slug": "ccc-chc-program",
-      "description": "High-need child health interventions for children under 5",
-      "budget": 500000,
-      "currency": "USD",
-      "start_date": "2024-01-01",
-      "end_date": "2024-12-31",
-      "organization": 1
-    }
-  },
-  {
-    "model": "solicitations.solicitation",
-    "pk": 1,
-    "fields": {
-      "title": "CommCare Connect Child Health Campaign (CCC-CHC) - Expression of Interest",
-      "description": "We are seeking partners to deliver door-to-door child health services including vitamin A supplementation, deworming tablets distribution, malnutrition screening, vaccine status checking, and ORS/Zinc drop-off for diarrhea management.",
-      "target_population": "Children under 5 in high-need regions",
-      "scope_of_work": "Door-to-door delivery of vitamin A, deworming tablets, malnutrition screening using MUAC measurements, vaccine checking and referrals, ORS/Zinc distribution for diarrhea cases, real-time data entry via CommCare Connect smartphone app",
-      "solicitation_type": "eoi",
-      "status": "completed",
-      "is_publicly_listed": true,
-      "program": 1,
-      "created_by": 1,
-      "estimated_scale": "40,000 children across multiple regions",
-      "expected_start_date": "2025-04-01",
-      "expected_end_date": "2025-07-31",
-      "application_deadline": "2025-02-15",
-      "date_created": "2024-01-01T00:00:00Z"
-    }
-  }
-]
